Revert "issue-1482: Minor package refactoring and cleanup"
This reverts commit dfea7e19060c928e9be1263ddfc294097cdf175d.
