+++ strongbox-storage/strongbox-storage-layout-providers/strongbox-storage-maven-layout/strongbox-maven-metadata-api/src/main/java/org/carlspring/strongbox/storage/metadata/VersionCollectionRequest.java
@@ -1,6 +1,6 @@
 package org.carlspring.strongbox.storage.metadata;
 
-import org.carlspring.strongbox.storage.metadata.maven.MetadataVersion;
+import org.carlspring.strongbox.storage.metadata.versions.MetadataVersion;
 
 import java.nio.file.Path;
 import java.util.ArrayList;
