+++ strongbox-storage/strongbox-storage-layout-providers/strongbox-storage-maven-layout/strongbox-maven-metadata-api/src/main/java/org/carlspring/strongbox/storage/metadata/VersionCollector.java
@@ -1,10 +1,10 @@
 package org.carlspring.strongbox.storage.metadata;
 
-import org.carlspring.strongbox.storage.metadata.maven.ArtifactVersionDirectoryFilter;
-import org.carlspring.strongbox.storage.metadata.maven.MetadataVersionComparator;
-import org.carlspring.strongbox.storage.metadata.maven.SnapshotVersionComparator;
-import org.carlspring.strongbox.storage.metadata.maven.MetadataVersion;
-import org.carlspring.strongbox.storage.metadata.maven.ArtifactVersionDirectoryVisitor;
+import org.carlspring.strongbox.io.filters.ArtifactVersionDirectoryFilter;
+import org.carlspring.strongbox.storage.metadata.comparators.MetadataVersionComparator;
+import org.carlspring.strongbox.storage.metadata.comparators.SnapshotVersionComparator;
+import org.carlspring.strongbox.storage.metadata.versions.MetadataVersion;
+import org.carlspring.strongbox.storage.metadata.visitors.ArtifactVersionDirectoryVisitor;
 
 import java.io.FileReader;
 import java.io.IOException;
